# service/Market/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Zifeldev_marketback_service_Market_internal_models.CreateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      price:
        type: number
      sizes:
        items:
          type: string
        type: array
      stock:
        minimum: 0
        type: integer
      title:
        type: string
    required:
    - category_id
    - price
    - stock
    - title
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.CreateSellerRequest:
    properties:
      description:
        type: string
      shop_name:
        type: string
    required:
    - shop_name
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      price:
        type: number
      seller_id:
        type: integer
      sizes:
        items:
          type: string
        type: array
      status:
        type: string
      stock:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.Seller:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      rating:
        type: number
      shop_name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.UpdateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      price:
        type: number
      sizes:
        items:
          type: string
        type: array
      status:
        type: string
      stock:
        type: integer
      title:
        type: string
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.UpdateSellerRequest:
    properties:
      description:
        type: string
      shop_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Marketplace API with products, categories, cart, orders, seller and
    admin management
  title: Market Service API
  version: "1.0"
paths:
...
  /api/seller/products:
    get:
      consumes:
      - application/json
      description: Get all products for current seller
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Product'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get seller products
      tags:
      - seller
    post:
      consumes:
      - application/json
      description: Create a new product for seller
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - seller
  /api/seller/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete seller's product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - seller
    put:
      consumes:
      - application/json
      description: Update seller's product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - seller
  /api/seller/profile:
    get:
      consumes:
      - application/json
      description: Get current user's seller profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Seller'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get seller profile
      tags:
      - seller
    put:
      consumes:
      - application/json
      description: Update current user's seller profile
      parameters:
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.UpdateSellerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Seller'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update seller profile
      tags:
      - seller
  /api/seller/register:
    post:
      consumes:
      - application/json
      description: Create a seller profile for the authenticated user
      parameters:
      - description: Seller data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.CreateSellerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Seller'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Register seller profile
      tags:
      - seller
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"